Migrated to TypeScript and renamed files
