--- spots-app-react/src/components/Footer.jsx
@@ -1,11 +1,0 @@
-function Footer() {
-  return (
-    <footer>
-      <div className="footer">
-        <p>2025 &copy; Spots App. #Tinyuka, AltSchool</p>
-      </div>
-    </footer>
-  );
-}
-
-export default Footer;
