--- spots-app-react/src/components/Header.jsx
@@ -1,9 +1,0 @@
-function Header() {
-  return (
-    <div className="header">
-      <img src="/circle17icons/logo/Logo.svg" alt="Spots Logo" />
-    </div>
-  );
-}
-
-export default Header;
