+++ spots-app-react/src/components/Header.tsx
@@ -1,0 +1,9 @@
+function Header() {
+  return (
+    <div className="header">
+      <img src="/circle17icons/logo/Logo.svg" alt="Spots Logo" />
+    </div>
+  );
+}
+
+export default Header;
