--- spots-app-react/src/components/PostContainer.jsx
@@ -1,25 +1,0 @@
-import PostCard from './PostCard';
-
-function PostsContainer({ posts, setPosts, onPreview }) {
-  const handleLike = (index) => {
-    const updatedPosts = [...posts];
-    updatedPosts[index].liked = !updatedPosts[index].liked;
-    setPosts(updatedPosts);
-  };
-
-  return (
-    <div className="posts-container">
-      {posts.map((post, index) => (
-        <PostCard
-          key={index}
-          post={post}
-          index={index}
-          onLike={handleLike}
-          onPreview={onPreview}
-        />
-      ))}
-    </div>
-  );
-}
-
-export default PostsContainer;
